Added speed parameter to moves.
Turn duration should always be duration 620 and speed 20 for 90° turn.
Turn duration should always be duration 1240 and speed 20 for 180° turn.
Speed mustn't be higher than 127. (Not even sure if drone can actually
go that fast).
Duration must NEVER be 0 for either type of movement. (It'll skip things
and do a mess.)